car_logic/requirements.txt
RPi.GPIO
adafruit-circuitpython-bmp280